22/11/22 23:39

+++ src/main/java/MyFirstProject/member/mapper/MemberMapper.java
@@ -15,7 +15,7 @@ public interface MemberMapper {
 
     MemberDto findMemberById(String loginId);
 
-    MemberDto findMemberByIdx(int memberIdx);
+    MemberDto findMemberByIdx(long memberIdx);
 
     int insertMember(JoinFormDto joinFormDto);
 }
